+++ eslint.config.js
@@ -1,9 +1,3 @@
-import { ignoreFile } from '@shgysk8zer0/eslint-config/ignoreFile.js';
-import browser from '@shgysk8zer0/eslint-config/browser.js';
+import { browser } from '@shgysk8zer0/eslint-config';
 
-export default [
-	ignoreFile,
-	browser({
-		ignores: ['**/*.min.js', '**/*.cjs', 'assets/dedent.js','assets/url-pattern.js'],
-	}),
-];
+export default browser({ ignores: ['**/*.min.js', '**/*.cjs', 'assets/dedent.js','assets/url-pattern.js'] });