+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [v0.4.4] - 2024-10-09
+
+### Added
+- Add `Error.isError` (imperfect, but works)
+
+### Fixed
+- Fixed and simplicy `Array.prototype.uniqueBy`
+
 ## [v0.4.3] - 2024-09-17
 
 ### Added